Refactored each module to set itself up
This means that each module finds its relevant DOM elements. This is not as easily testable or
injectible, but I felt too much like rewriting React-style code... So I'll maintain this simpler style
and throw errors if any of the elements are missing.
 # Please enter the commit message for your changes.
Lines starting # with '#' will be ignored, and an empty message aborts the commit.